+++ plugins/domains/pastie.org.js
@@ -10,8 +10,8 @@ module.exports = {
 
     getLink: function(urlMatch) {
         return {
-            href: "http://pastie.org/" + urlMatch[1] +".js",
-            type: CONFIG.T.javascript, 
+            html: '<script src="http://pastie.org/' + urlMatch[1] + '.js" type="text/javascript"></script>',
+            type: CONFIG.T.text_html,
             rel: CONFIG.R.reader // not inline :\
         };
     },