+++ plugins/domains/rapgenius.com.js
@@ -1,6 +1,6 @@
 module.exports = {
 
-    re: /^https?:\/\/(?:[\w\-]+\.)?genius\.com\/(?!jobs)([a-z-]+)/i,
+    re: /^https?:\/\/(?:[\w\-]+\.)?genius\.com\/(?!jobs)([a-zA-Z\-]+)/i,
 
     mixins: [
         "favicon",
@@ -14,7 +14,8 @@ module.exports = {
     getLink: function(urlMatch, twitter) {
         var id = twitter.app.url.iphone.match(/\d+/)[0];
         return {
-            html: '<div id="rg_embed_link_' + id + '" class="rg_embed_link">Read <a href="http://rapgenius.com/D12-my-band-lyrics">' + twitter.title + '</a> on Genius</div><script src="//' + urlMatch[1] + '.genius.com/songs/' + id + '/embed.js?dark=1"></script>',
+            html: '<div id="rg_embed_link_' + id + '" class="rg_embed_link" data-song-id="' + id + '"></div><script src="//genius.com/songs/' + id + '/embed.js?dark=1"></script>',
+
             type: CONFIG.T.text_html,
             rel: [CONFIG.R.reader, CONFIG.R.ssl]
         };